requirements.txt
pint
numpy
scipy
streamlit 
pyyaml 
pandas
h5py
plotly